requirements.txt
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
google-cloud-language==1.3.0
gunicorn==20.0.0
itsdangerous==1.1.0
pytz==2019.3
structlog==19.2.0
termcolor==1.1.0
WTForms==2.2.1
